// src/app/core/models/masters/human-capital/unit.ts
export interface Unit {
    id: number;
    title: string;
    status: string;
    status_id :number;
    status_label: string;
}

export interface UnitPaginationResponse {
    current_page: number;
    total: number;
    per_page: number;
    last_page: number;
    from: number;
    data: Unit[];
}